+++ src/controllers/users/userController.js
@@ -1,17 +1,18 @@
-import { findUserById } from "../services/userService.js";
-import { updateUserSettingById } from "../services/userSettingService.js";
-import { stringToSnakeCase } from "../utils/caseConverter.js";
+import {
+  registerUser,
+  findUserById,
+} from "../../services/users/userService.js";
+import { stringToSnakeCase } from "../../utils/caseConverter.js";
 
-export const updateUserSettings = async (req, res, next) => {
+export const createUser = async (req, res, next) => {
   try {
-    const { userId } = req.params;
     const reqSettingFields = ["isAdDetect", "isReturnChannel"];
     const settingFields = reqSettingFields.map((key) => [
       key,
       stringToSnakeCase(key),
     ]);
 
-    const updateFields = {};
+    const insertFields = {};
     for (const [reqKey, dbKey] of settingFields) {
       const value = req.body?.[reqKey];
       if (value !== undefined) {
@@ -21,26 +22,29 @@ export const updateUserSettings = async (req, res, next) => {
             error: "올바르지 않은 형식입니다.",
           });
         }
-        updateFields[dbKey] = value;
+        insertFields[dbKey] = value;
       }
     }
 
-    if (Object.keys(updateFields).length === 0) {
-      return res.status(400).json({
-        status: 400,
-        error: "올바르지 않은 형식입니다.",
-      });
-    }
+    const message = await registerUser(insertFields);
+    res.status(201).json(message);
+  } catch (error) {
+    next(error);
+  }
+};
 
+export const getUserById = async (req, res, next) => {
+  try {
+    const { userId } = req.params;
     const user = await findUserById(Number(userId));
+
     if (!user) {
       return res
         .status(404)
         .json({ status: 404, error: "사용자를 찾을 수 없습니다." });
     }
 
-    const message = await updateUserSettingById(Number(userId), updateFields);
-    res.status(200).json(message);
+    res.status(200).json(user);
   } catch (error) {
     next(error);
   }
