+++ src/routes/reports/index.js
@@ -1,6 +1,6 @@
 import express from "express";
 
-import { postAdReport } from "../../controllers/reportController.js";
+import { postAdReport } from "../../controllers/reports/reportController.js";
 import { validateReportBody } from "../../validators/reportValidator.js";
 
 const router = express.Router();
